make first domain test pass with new meteor testing
