+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.3-beta.12
+METEOR@1.3-rc.7
