+++ .meteor/versions
@@ -1,73 +1,71 @@
-allow-deny@1.0.1-beta.12
-autoupdate@1.2.5-beta.12
-avital:mocha@2.1.0_8
-babel-compiler@6.5.0-beta.12
-babel-runtime@0.1.5-beta.12
-base64@1.0.5-beta.12
-binary-heap@1.0.5-beta.12
-blaze@2.1.4-beta.12
-blaze-html-templates@1.0.1
-blaze-tools@1.0.5-beta.12
-boilerplate-generator@1.0.5-beta.12
-caching-compiler@1.0.1-beta.12
-caching-html-compiler@1.0.3-beta.12
-callback-hook@1.0.5-beta.12
-check@1.1.1-beta.12
-coffeescript@1.0.12-beta.12
+allow-deny@1.0.2-rc.7
+autoupdate@1.2.6-rc.7
+babel-compiler@6.5.2-rc.7
+babel-runtime@0.1.6-rc.7
+base64@1.0.6-rc.7
+binary-heap@1.0.6-rc.7
+blaze@2.1.5-rc.7
+blaze-html-templates@1.0.2-rc.7
+blaze-tools@1.0.6-rc.7
+boilerplate-generator@1.0.6-rc.7
+caching-compiler@1.0.2-rc.7
+caching-html-compiler@1.0.4-rc.7
+callback-hook@1.0.6-rc.7
+check@1.1.2-rc.7
+coffeescript@1.0.15-rc.7
 cosmos:browserify@0.9.4
-ddp@1.2.2
-ddp-client@1.2.2-beta.12
-ddp-common@1.2.2
-ddp-server@1.2.3-beta.12
-deps@1.0.9
-diff-sequence@1.0.2-beta.12
-ecmascript@0.4.0-beta.12
-ecmascript-runtime@0.2.7-beta.12
-ejson@1.0.8-beta.12
-es5-shim@4.3.2-beta.12
-fastclick@1.0.8-beta.12
+ddp@1.2.3-rc.7
+ddp-client@1.2.3-rc.7
+ddp-common@1.2.3-rc.7
+ddp-server@1.2.4-rc.7
+deps@1.0.10-rc.7
+diff-sequence@1.0.3-rc.7
+ecmascript@0.4.1-rc.7
+ecmascript-runtime@0.2.8-rc.7
+ejson@1.0.9-rc.7
+es5-shim@4.5.8-rc.7
+fastclick@1.0.9-rc.7
 fongandrew:find-and-modify@0.2.2
-geojson-utils@1.0.5-beta.12
-hot-code-push@1.0.1-beta.12
-html-tools@1.0.6-beta.12
-htmljs@1.0.6-beta.12
-http@1.1.2-beta.12
-id-map@1.0.4
-jquery@1.11.5-beta.12
+geojson-utils@1.0.6-rc.7
+hot-code-push@1.0.2-rc.7
+html-tools@1.0.7-rc.7
+htmljs@1.0.7-rc.7
+http@1.1.3-rc.7
+id-map@1.0.5-rc.7
+jquery@1.11.6-rc.7
 kadira:flow-router@2.10.1
-launch-screen@1.0.6-beta.12
-livedata@1.0.15
-logging@1.0.9-beta.12
-meteor@1.1.11-beta.12
-meteor-base@1.0.1
-meteor-env-dev@0.0.2-beta.12
-meteor-env-prod@0.0.2-beta.12
+launch-screen@1.0.8-rc.7
+livedata@1.0.16-rc.7
+logging@1.0.10-rc.7
+meteor@1.1.12-rc.7
+meteor-base@1.0.2-rc.7
 mikowals:batch-insert@1.1.13
-minifier-css@1.1.8-beta.12
-minifier-js@1.1.8-beta.12
-minimongo@1.0.11-beta.12
-mobile-experience@1.0.1
-mobile-status-bar@1.0.8-beta.12
-modules@0.5.0-beta.12
-modules-runtime@0.5.0-beta.12
-mongo@1.1.4-beta.12
-mongo-id@1.0.1
-npm-mongo@1.4.40-beta.12
-observe-sequence@1.0.8-beta.12
-ordered-dict@1.0.4
+minifier-css@1.1.9-rc.7
+minifier-js@1.1.9-rc.7
+minimongo@1.0.12-rc.7
+mobile-experience@1.0.2-rc.7
+mobile-status-bar@1.0.10-rc.7
+modules@0.5.1-rc.7
+modules-runtime@0.6.1-rc.7
+mongo@1.1.5-rc.7
+mongo-id@1.0.2-rc.7
+npm-mongo@1.4.41-rc.7
+observe-sequence@1.0.9-rc.7
+ordered-dict@1.0.5-rc.7
 practicalmeteor:chai@2.1.0_1
 practicalmeteor:loglevel@1.2.0_2
+practicalmeteor:mocha@2.1.1-rc.1
 practicalmeteor:mocha-core@0.1.4
 practicalmeteor:munit@2.1.5
 practicalmeteor:sinon@1.14.1_2
-promise@0.5.2-beta.12
-random@1.0.6-beta.12
-reactive-dict@1.1.4-beta.12
-reactive-var@1.0.6
-reload@1.1.5-beta.12
-retry@1.0.4
-routepolicy@1.0.7-beta.12
-session@1.1.2-beta.12
+promise@0.6.4-rc.7
+random@1.0.7-rc.7
+reactive-dict@1.1.5-rc.7
+reactive-var@1.0.7-rc.7
+reload@1.1.6-rc.7
+retry@1.0.5-rc.7
+routepolicy@1.0.8-rc.7
+session@1.1.3-rc.7
 space:base@4.1.2
 space:domain@0.2.1
 space:event-sourcing@3.0.1
@@ -78,17 +76,18 @@ space:testing-messaging@3.0.1
 space:ui@6.0.0
 space:vo-monetary@0.2.0
 space:vo-user@0.3.0
-spacebars@1.0.8-beta.12
-spacebars-compiler@1.0.8-beta.12
-standard-minifier-css@1.0.3-beta.12
-standard-minifier-js@1.0.3-beta.12
-templating@1.1.6-beta.12
-templating-tools@1.0.1-beta.12
-test-helpers@1.0.6-beta.12
-tinytest@1.0.7-beta.12
-tracker@1.0.10-beta.12
-ui@1.0.8
-underscore@1.0.5-beta.12
-url@1.0.6-beta.12
-webapp@1.2.5-beta.12
-webapp-hashing@1.0.6-beta.12
+spacebars@1.0.9-rc.7
+spacebars-compiler@1.0.9-rc.7
+standard-minifier-css@1.0.4-rc.7
+standard-minifier-js@1.0.4-rc.7
+templating@1.1.7-rc.7
+templating-tools@1.0.2-rc.7
+test-helpers@1.0.7-rc.7
+tinytest@1.0.8-rc.7
+tmeasday:test-reporter-helpers@0.2.1
+tracker@1.0.11-rc.7
+ui@1.0.9-rc.7
+underscore@1.0.6-rc.7
+url@1.0.7-rc.7
+webapp@1.2.6-rc.7
+webapp-hashing@1.0.7-rc.7
