+++ packages/.gitignore
@@ -1,3 +1,4 @@
 space-messaging
 space-event-sourcing
 space-testing
+space-base
