+++ source/client/index.js
@@ -1,3 +1,2 @@
-import App from '/source/imports/ui/application.js'
-
+import App from '/source/imports/ui/application.js';
 App.start();
