+++ source/imports/domain/commands.js
@@ -1,0 +1,12 @@
+import Contact from './value-objects/contact';
+
+export default Space.messaging.define(Space.domain.Command, {
+
+  OpenBankAccount: {
+    owner: Contact
+  },
+
+  CloseBankAccount: {
+  }
+
+});
