+++ source/imports/domain/events.js
@@ -1,0 +1,9 @@
+import Contact from './value-objects/contact';
+
+export default Space.messaging.define(Space.domain.Event,  {
+
+  BankAccountOpened: {
+    owner: Contact,
+    balance: Money
+  }
+});
