--- source/imports/es-routers/bank-account-router.js
@@ -1,15 +1,0 @@
-import BankAccount from '../domain/bank-accounts/bank-account';
-import { OpenBankAccount, CloseBankAccount } from '../domain-commands/commands';
-
-const BankAccountRouter = Space.eventSourcing.Router.extend('BankAccountRouter', {
-
-  eventSourceable: BankAccount,
-  initializingMessage: OpenBankAccount,
-
-  routeCommands: [
-    this.CloseBankAccount
-  ]
-
-});
-
-export default BankAccountRouter;
